Set render in board controller
